+++ public/results.php
@@ -1,50 +1,78 @@
 <?php
+// display results
 if (count($results[$p]) > 0) {
+	//print_r($_SERVER);
 ?>
 <div class="container-fluid searchresults">
   <div class="row">
-    <div class="alert alert-dismissible alert-success col-xs-4">
-      <button type="button" class="close" data-dismiss="alert">&times;</button>
-      <span class="glyphicon glyphicon-search"></span> <strong><?php echo $total; ?> files found</strong>.
-    </div>
+		<div class="col-xs-6">
+			<div class="row">
+				<div class="alert alert-dismissible alert-success col-xs-8">
+					<button type="button" class="close" data-dismiss="alert">&times;</button>
+					<span class="glyphicon glyphicon-search"></span> <strong><?php echo $total; ?> files found</strong>.
+				</div>
+			</div>
+		</div>
+		<div class="col-xs-6">
+			<div class="row">
+				<div class="alert alert-dismissible alert-warning col-xs-8 pull-right unsavedChangesAlert" style="display:none;">
+					<button type="button" class="close" data-dismiss="alert">&times;</button>
+					<span class="glyphicon glyphicon-save"></span> <strong> <span class="changetagcounter"></span>. Press Tag files to save changes.</strong>
+				</div>
+			</div>
+		</div>
   </div>
   <div class="row">
     <form method="post" action="/tagfiles.php" class="form-inline">
-      <div class="form-group pull-right">
-        <input type="text" class="search form-control" placeholder="Search within results">
-      </div>
-    <span class="counter pull-right"></span>
-    <table class="table table-striped table-hover results" style="word-break:break-word;word-wrap:break-word;">
+			<div class="col-xs-12 text-right">
+				<div class="row">
+				<div class="btn-group">
+					<button class="btn btn-default tagAllDelete" type="button" name="tagAll"> Select All Delete</button>
+					<button class="btn btn-default tagAllArchive" type="button" name="tagAll"> All Archive</button>
+					<button class="btn btn-default tagAllKeep" type="button" name="tagAll"> All Keep</button>
+					<button class="btn btn-default tagAllUntagged" type="button" name="tagAll"> All Untagged</button>
+				</div>
+				<button type="button" class="btn btn-default reload-results"><i class="glyphicon glyphicon-refresh"></i> Reload</button>
+				<button type="submit" class="btn btn-default button-tagfiles"><i class="glyphicon glyphicon-tag"></i> Tag files</button>
+				<div class="form-group">
+					<input type="text" class="search form-control" placeholder="Search within results">
+				</div>
+				</div>
+		</div>
+    <div class="counter pull-right"></div>
+    <table class="table table-striped table-hover results">
       <thead>
         <tr>
           <th class="text-nowrap">#</th>
-          <th class="text-nowrap">Filename</th>
-          <th class="text-nowrap">Parent Path</th>
-          <th class="text-nowrap">Size</th>
-          <th class="text-nowrap">Owner</th>
-          <th class="text-nowrap">Group</th>
-          <th class="text-nowrap">Last Modified (utc)</th>
-          <th class="text-nowrap">Last Access (utc)</th>
-          <th class="text-nowrap">Tag (del/arch/keep)</th>
-          <th class="text-nowrap">Custom Tag</th>
+          <th class="text-nowrap">Filename <a href="<?php echo $_SERVER["REQUEST_URI"] . '&sort=filename&sortorder=asc' ?>"><i class="glyphicon glyphicon-chevron-up sortarrows"></i></a><a href="<?php echo $_SERVER["REQUEST_URI"] . '&sort=filename&sortorder=desc' ?>"><i class="glyphicon glyphicon-chevron-down sortarrows"></i></a></th>
+          <th class="text-nowrap">Parent Path <a href="<?php echo $_SERVER["REQUEST_URI"] . '&sort=path_parent&sortorder=asc' ?>"><i class="glyphicon glyphicon-chevron-up sortarrows"></i></a><a href="<?php echo $_SERVER["REQUEST_URI"] . '&sort=path_parent&sortorder=desc' ?>"><i class="glyphicon glyphicon-chevron-down sortarrows"></i></a></th>
+					<th class="text-nowrap">Size <a href="<?php echo $_SERVER["REQUEST_URI"] . '&sort=filesize&sortorder=asc' ?>"><i class="glyphicon glyphicon-chevron-up sortarrows"></i></a><a href="<?php echo $_SERVER["REQUEST_URI"] . '&sort=filesize&sortorder=desc' ?>"><i class="glyphicon glyphicon-chevron-down sortarrows"></i></a></th>
+          <th class="text-nowrap">Owner <a href="<?php echo $_SERVER["REQUEST_URI"] . '&sort=owner&sortorder=asc' ?>"><i class="glyphicon glyphicon-chevron-up sortarrows"></i></a><a href="<?php echo $_SERVER["REQUEST_URI"] . '&sort=owner&sortorder=desc' ?>"><i class="glyphicon glyphicon-chevron-down sortarrows"></i></a></th>
+          <th class="text-nowrap">Group <a href="<?php echo $_SERVER["REQUEST_URI"] . '&sort=group&sortorder=asc' ?>"><i class="glyphicon glyphicon-chevron-up sortarrows"></i></a><a href="<?php echo $_SERVER["REQUEST_URI"] . '&sort=group&sortorder=desc' ?>"><i class="glyphicon glyphicon-chevron-down sortarrows"></i></a></th>
+          <th class="text-nowrap">Last Modified (utc) <a href="<?php echo $_SERVER["REQUEST_URI"] . '&sort=last_modified&sortorder=asc' ?>"><i class="glyphicon glyphicon-chevron-up sortarrows"></i></a><a href="<?php echo $_SERVER["REQUEST_URI"] . '&sort=last_modified&sortorder=desc' ?>"><i class="glyphicon glyphicon-chevron-down sortarrows"></i></a></th>
+          <th class="text-nowrap">Last Access (utc) <a href="<?php echo $_SERVER["REQUEST_URI"] . '&sort=last_access&sortorder=asc' ?>"><i class="glyphicon glyphicon-chevron-up sortarrows"></i></a><a href="<?php echo $_SERVER["REQUEST_URI"] . '&sort=last_access&sortorder=desc' ?>"><i class="glyphicon glyphicon-chevron-down sortarrows"></i></a></th>
+          <th class="text-nowrap">Tag (del/arch/keep) <a href="<?php echo $_SERVER["REQUEST_URI"] . '&sort=tag&sortorder=asc' ?>"><i class="glyphicon glyphicon-chevron-up sortarrows"></i></a><a href="<?php echo $_SERVER["REQUEST_URI"] . '&sort=tag&sortorder=desc' ?>"><i class="glyphicon glyphicon-chevron-down sortarrows"></i></a></th>
+          <th class="text-nowrap">Custom Tag <a href="<?php echo $_SERVER["REQUEST_URI"] . '&sort=tag_custom&sortorder=asc' ?>"><i class="glyphicon glyphicon-chevron-up sortarrows"></i></a><a href="<?php echo $_SERVER["REQUEST_URI"] . '&sort=tag_custom&sortorder=desc' ?>"><i class="glyphicon glyphicon-chevron-down sortarrows"></i></a></th>
         </tr>
         <tr class="warning no-result">
           <td colspan="10"><i class="fa fa-warning"></i> No result</td>
         </tr>
       </thead>
       <tfoot>
-        <th>#</th>
-        <th>Filename</th>
-        <th>Parent Path</th>
-        <th>Size</th>
-        <th>Owner</th>
-        <th>Group</th>
-        <th>Last Modified (utc)</th>
-        <th>Last Access (utc)</th>
-        <th>Tag (del/arch/keep)</th>
-        <th>Custom Tag</th>
+				<tr>
+					<th class="text-nowrap">#</th>
+					<th class="text-nowrap">Filename</th>
+					<th class="text-nowrap">Parent Path</th>
+					<th class="text-nowrap">Size</th>
+					<th class="text-nowrap">Owner</th>
+					<th class="text-nowrap">Group</th>
+					<th class="text-nowrap">Last Modified (utc)</th>
+					<th class="text-nowrap">Last Access (utc)</th>
+					<th class="text-nowrap">Tag (del/arch/keep)</th>
+					<th class="text-nowrap">Custom Tag</th>
+				</tr>
       </tfoot>
-      <tbody>
+      <tbody id="results-tbody">
       <?php
         error_reporting(E_ALL ^ E_NOTICE);
         $limit = $searchParams['size'];
@@ -63,7 +91,7 @@ if (count($results[$p]) > 0) {
         <td class="text-nowrap"><?php echo $file['group']; ?></td>
         <td class="text-nowrap"><?php echo $file['last_modified']; ?></td>
         <td class="text-nowrap"><?php echo $file['last_access']; ?></td>
-        <td class="text-nowrap"><div class="btn-group" style="white-space:nowrap;" data-toggle="buttons">
+        <td class="text-nowrap"><div class="btn-group tagButtons" style="white-space:nowrap;" data-toggle="buttons">
             <label class="tagDeleteLabel btn btn-warning btn-sm <?php if ($file['tag'] == 'delete') { echo 'active'; }?>" style="display:inline-block;float:none;" id="highlightRowDelete">
               <input class="tagDeleteInput" type="radio" name="ids_tag[<?php echo $result['_id']; ?>]" value="delete" <?php if ($file['tag'] == 'delete') { echo 'checked'; }; ?> /><span class="glyphicon glyphicon-trash"></span>
             </label>
@@ -74,10 +102,10 @@ if (count($results[$p]) > 0) {
               <input class="tagKeepInput" type="radio" name="ids_tag[<?php echo $result['_id']; ?>]" value="keep" <?php if ($file['tag'] == 'keep') { echo 'checked'; }; ?> /><span class="glyphicon glyphicon-floppy-saved"></span>
             </label>
 						<label class="tagUntaggedLabel btn btn-default btn-sm" style="display:inline-block;float:none;" id="highlightRowUntagged">
-              <input class="tagUntaggedInput" type="radio" name="ids_tag[<?php echo $result['_id']; ?>]" value="untagged" /><span class="glyphicon glyphicon-remove-sign"></span>
+              <input class="tagUntaggedInput" type="radio" name="ids_tag[<?php echo $result['_id']; ?>]" value="untagged" <?php if ($file['tag'] == 'untagged') { echo 'checked'; }; ?> /><span class="glyphicon glyphicon-remove-sign"></span>
             </label>
           </div></td>
-        <td class="text-nowrap"><input type="text" name="ids_tag_custom[<?php echo $result['_id']; ?>]" value="<?php echo $file['tag_custom']; ?>" ?></td>
+        <td class="text-nowrap"><input type="text" name="ids_tag_custom[<?php echo $result['_id']; ?>]" class="custom-tag-input" value="<?php echo $file['tag_custom']; ?>"></td>
       </tr>
       <?php
         } // END foreach loop over results
@@ -85,22 +113,24 @@ if (count($results[$p]) > 0) {
       </tbody>
     </table>
   </div>
-  <div class="row pull-right">
-    <div class="col-xs-12">
-      <p class="text-right">
+  <div class="row">
+    <div class="col-xs-12 text-right">
+			<div class="row">
       <div class="btn-group">
-        <button class="btn btn-default" type="button" name="tagAll" id="tagAllDelete" /> Select All Delete</button>
-        <button class="btn btn-default" type="button" name="tagAll" id="tagAllArchive" /> Select All Archive</button>
-        <button class="btn btn-default" type="button" name="tagAll" id="tagAllKeep" /> Select All Keep</button>
+        <button class="btn btn-default tagAllDelete" type="button" name="tagAll"> Select All Delete</button>
+        <button class="btn btn-default tagAllArchive" type="button" name="tagAll"> All Archive</button>
+        <button class="btn btn-default tagAllKeep" type="button" name="tagAll"> All Keep</button>
+				<button class="btn btn-default" type="button" name="tagAll"> All Untagged</button>
       </div>
-      <button type="button" id="refresh" class="btn btn-default">Refresh</button>
-      <button type="submit" class="btn btn-primary">Tag files</button>
-      </p>
+      <button type="button" class="btn btn-default reload-results"><i class="glyphicon glyphicon-refresh"></i> Reload</button>
+      <button type="submit" class="btn btn-default button-tagfiles"><i class="glyphicon glyphicon-tag"></i> Tag files</button>
       </form>
+			</div>
     </div>
   </div>
   <div class="row">
     <div class="col-xs-12 text-right">
+			<div class="row">
       <?php
       // pagination
       if ($total > $limit) {
@@ -147,6 +177,7 @@ if (count($results[$p]) > 0) {
       </ul>
       <?php } ?>
     </div>
+	</div>
   </div>
 </div>
 <?php